<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation
In Triton, cache_modifier is used to specify how the memory loading cache strategy should be handled. .cg (cache) is a memory access optimization option provided by Triton. This PR applies it to enable caching during data loading, allowing the loaded data to better utilize the GPU's cache mechanism, thereby improving performance.

The benchmark result shows in Triton Attention (wo setting window_size, WINDOW_SIZE=-1 in table below) it gains 17-20% speedup.

Main:
$python ./benchmark/kernels/sliding_window_attention_triton/bench_triton_swa_kernel.py
extend_attention_triton_vs_torch:
     N_CTX  WINDOW_SIZE       torch      triton
0   1024.0         -1.0   13.313979    3.400983
1   1024.0        127.0   13.998181    2.839544
2   1024.0        256.0   13.671337    3.277899
3   1024.0        512.0   13.484036    3.724898
4   2048.0         -1.0   37.498367   12.688069
5   2048.0        127.0   39.144402    8.056528
6   2048.0        256.0   38.922625    9.115718
7   2048.0        512.0   38.468863   10.897383
8   4096.0         -1.0  136.244034   49.277617
9   4096.0        127.0  141.462463   24.554945
10  4096.0        256.0  143.137283   26.608192
11  4096.0        512.0  140.915131   30.392427
12  8192.0         -1.0  404.721649  185.647873
13  8192.0        127.0  423.228149   79.989822
14  8192.0        256.0  421.111450   84.352928
15  8192.0        512.0  419.148163   92.420738

============================================================

This PR:
$python ./benchmark/kernels/sliding_window_attention_triton/bench_triton_swa_kernel.py
extend_attention_triton_vs_torch:
     N_CTX  WINDOW_SIZE       torch      triton
0   1024.0         -1.0   13.216082    2.986481
1   1024.0        127.0   13.694473    2.850151
2   1024.0        256.0   13.617353    3.294159
3   1024.0        512.0   13.488297    3.724024
4   2048.0         -1.0   37.403521   10.948128
5   2048.0        127.0   39.125616    8.041440
6   2048.0        256.0   39.439072    9.110515
7   2048.0        512.0   38.453457   10.890365
8   4096.0         -1.0  136.164291   41.194319
9   4096.0        127.0  141.763000   24.516776
10  4096.0        256.0  140.950470   26.599690
11  4096.0        512.0  139.955719   30.417290
12  8192.0         -1.0  403.730072  148.042236
13  8192.0        127.0  421.817413   79.692383
14  8192.0        256.0  420.936523   84.381027
15  8192.0        512.0  419.162079   92.444572


<!-- Describe the purpose and goals of this pull request. -->

## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->

## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).